+++ .idea/.gitignore
@@ -2,7 +2,7 @@
 /shelf/
 /workspace.xml
 # Datasource local storage ignored files
-/../../../../../../../../:\Users\anton\Documents\SkCourses\ML\AEproject\.idea/dataSources/
+/../../../../../../../:\Users\User_Anton\Documents\SkCourses\MLproject\.idea/dataSources/
 /dataSources.local.xml
 # Editor-based HTTP Client requests
 /httpRequests/
