--- dnn-mode-connectivity/models/__init__.py
@@ -1,5 +1,0 @@
-from .convfc import *
-from .vgg import *
-from .preresnet import *
-from .wide_resnet import *
-from .autoencoder import *
