--- dnn-mode-connectivity/utils.py
@@ -1,170 +1,0 @@
-import numpy as np
-import os
-import torch
-import torch.nn.functional as F
-import curves
-
-
-def l2_regularizer(weight_decay):
-    def regularizer(model):
-        l2 = 0.0
-        for p in model.parameters():
-            l2 += torch.sqrt(torch.sum(p ** 2))
-        return 0.5 * weight_decay * l2
-    return regularizer
-
-
-def cyclic_learning_rate(epoch, cycle, alpha_1, alpha_2):
-    def schedule(it):
-        t = ((epoch % cycle) + it) / cycle
-        if t < 0.5:
-            return alpha_1 * (1.0 - 2.0 * t) + alpha_2 * 2.0 * t
-        else:
-            return alpha_1 * (2.0 * t - 1.0) + alpha_2 * (2.0 - 2.0 * t)
-    return schedule
-
-
-def adjust_learning_rate(optimizer, lr):
-    for param_group in optimizer.param_groups:
-        param_group['lr'] = lr
-    return lr
-
-
-def save_checkpoint(direc, epoch, name='checkpoint', **kwargs):
-    state = {
-        'epoch': epoch,
-    }
-    state.update(kwargs)
-    filepath = os.path.join(direc, '%s-%d.pt' % (name, epoch))
-    torch.save(state, filepath)
-
-
-def train(train_loader, model, optimizer, criterion,
-          device, regularizer=None, lr_schedule=None):
-    loss_sum = 0.0
-    correct = 0.0
-
-    num_iters = len(train_loader)
-    model.train()
-    for it, (inp, target) in enumerate(train_loader):
-        if lr_schedule is not None:
-            lr = lr_schedule(it / num_iters)
-            adjust_learning_rate(optimizer, lr)
-        inp = inp.to(device)
-        target = target.to(device)
-
-        output = model(input)
-        loss = criterion(output, target)
-        if regularizer is not None:
-            loss += regularizer(model)
-
-        optimizer.zero_grad()
-        loss.backward()
-        optimizer.step()
-
-        loss_sum += loss.item() * inp.size(0)
-        pred = output.data.argmax(1, keepdim=True)
-        correct += pred.eq(target.data.view_as(pred)).sum().item()
-
-    return {
-        'loss': loss_sum / len(train_loader.dataset),
-        'accuracy': correct * 100.0 / len(train_loader.dataset),
-    }
-
-
-def test(test_loader, model, criterion,
-         device, regularizer=None, **kwargs):
-    loss_sum = 0.0
-    nll_sum = 0.0
-    correct = 0.0
-
-    model.eval()
-
-    for inp, target in test_loader:
-        inp = inp.to(device)
-        target = target.to(device)
-
-        output = model(inp, **kwargs)
-        nll = criterion(output, target)
-        loss = nll.clone()
-        if regularizer is not None:
-            loss += regularizer(model)
-
-        nll_sum += nll.item() * inp.size(0)
-        loss_sum += loss.item() * inp.size(0)
-        pred = output.data.argmax(1, keepdim=True)
-        correct += pred.eq(target.data.view_as(pred)).sum().item()
-
-    return {
-        'nll': nll_sum / len(test_loader.dataset),
-        'loss': loss_sum / len(test_loader.dataset),
-        'accuracy': correct * 100.0 / len(test_loader.dataset),
-    }
-
-
-def predictions(test_loader, model, device,
-                **kwargs):
-    model.eval()
-    preds = []
-    targets = []
-    for inp, target in test_loader:
-        inp = inp.to(device)
-        output = model(input, **kwargs)
-        probs = F.softmax(output, dim=1)
-        preds.append(probs.cpu().data.numpy())
-        targets.append(target.numpy())
-    return np.vstack(preds), np.concatenate(targets)
-
-
-def isbatchnorm(module):
-    return issubclass(module.__class__, torch.nn.modules.batchnorm._BatchNorm) or \
-           issubclass(module.__class__, curves._BatchNorm)
-
-
-def _check_bn(module, flag):
-    if isbatchnorm(module):
-        flag[0] = True
-
-
-def check_bn(model):
-    flag = [False]
-    model.apply(lambda module: _check_bn(module, flag))
-    return flag[0]
-
-
-def reset_bn(module):
-    if isbatchnorm(module):
-        module.reset_running_stats()
-
-
-def _get_momenta(module, momenta):
-    if isbatchnorm(module):
-        momenta[module] = module.momentum
-
-
-def _set_momenta(module, momenta):
-    if isbatchnorm(module):
-        module.momentum = momenta[module]
-
-
-def update_bn(loader, model, device,
-              **kwargs):
-    if not check_bn(model):
-        return
-    model.train()
-    momenta = {}
-    model.apply(reset_bn)
-    model.apply(lambda module_: _get_momenta(module_, momenta))
-    num_samples = 0
-    for inp, _ in loader:
-        inp = inp.to(device)
-        batch_size = inp.data.size(0)
-
-        momentum = batch_size / (num_samples + batch_size)
-        for module_ in momenta.keys():
-            module_.momentum = momentum
-
-        model(inp, **kwargs)
-        num_samples += batch_size
-
-    model.apply(lambda module_: _set_momenta(module_, momenta))
