+++ requirements.txt
@@ -1,10 +1,8 @@
-numpy~=1.18.5
-torch
 gdown~=4.4.0
-natsort~=8.1.0
-pillow~=9.0.0
-torchvision~=0.11.1
-tqdm~=4.47.0
-requests~=2.24.0
-
-matplotlib~=3.2.2
+torch
+torchvision~=0.10.1
+natsort~=7.1.1
+pillow~=8.4.0
+numpy~=1.21.2
+scipy~=1.7.1
+tabulate~=0.8.9
